Moved Nuget Inspector installation into NugetInspectorInstaller Moved Nuget Inspector version resolution into NugetInspectorVersionResolver
